my-python-project/requirements.txt
PyPDF2
pytesseract
Pillow
regex
tkinter
logging